requirements_train.txt
# Training dependencies — only needed to train the model locally.
# Not required for running the backend server.
#
# Install with:
#   pip install -r requirements_train.txt
#
# Then download the dataset and run training:
#   python download_dataset.py
#   python train.py

torch>=2.1.0
torchvision>=0.16.0
onnx>=1.14.0
openimages>=0.3.0
Pillow>=10.0.0
